pyproject.toml
[project]
name = "mavica-tools"
version = "0.1.0"
description = "Floppy disk recovery and troubleshooting toolkit for Sony Mavica cameras"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["mavica", "floppy", "recovery", "jpeg", "sony", "camera", "retro"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: System :: Recovery Tools",
    "Topic :: Utilities",
]
dependencies = [
    "Pillow>=9.0",
    "textual>=0.50",
]

[project.optional-dependencies]
gps = [
    "piexif>=1.1.3",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "piexif>=1.1.3",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.urls]
Homepage = "https://github.com/gyng/mavica-tools"
Repository = "https://github.com/gyng/mavica-tools"
Issues = "https://github.com/gyng/mavica-tools/issues"

[project.scripts]
mavica = "mavica_tools.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]
ignore = [
    "E501",    # line length handled by formatter
    "E402",    # module-level import not at top — acceptable in CLI scripts
    "E701",    # multiple statements on one line — used sparingly
    "SIM105",  # suppressible-exception — contextlib.suppress not always clearer
    "SIM108",  # ternary operator — readability preference
    "SIM102",  # collapsible-if — nested ifs sometimes clearer with comments
    "SIM110",  # reimplemented-builtin — any() not always clearer
    "RUF012",  # mutable-class-default — Textual widgets use this pattern
    "RUF013",  # implicit-optional — keep explicit None defaults
    "UP007",   # X | Y union syntax — keep Optional for 3.10
    "B007",    # unused loop control variable — common in unpacking
    "B905",    # zip-without-explicit-strict — verbose for no safety gain here
    "RUF059",  # unused unpacked variable — common in tuple unpacking
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["F841"]  # unused variables common in test setup

[tool.ruff.lint.isort]
known-first-party = ["mavica_tools"]

[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = false
ignore_missing_imports = true
no_implicit_optional = false

[[tool.mypy.overrides]]
module = "mavica_tools.tui.*"
ignore_errors = true
